// src/app/services/api-services/models/employee-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EmployeeDto {
  dateOfBirth?: (string | null);
  departmentId?: (number | null);
  departmentName?: (string | null);
  email?: string;
  fullName?: string;
  id?: number;
  phone?: (string | null);
  position?: (string | null);
  userId?: (string | null);
}